% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_calc.R
\name{emis_calc}
\alias{emis_calc}
\title{Funkcja testowa na razie tylko mnozy dwa argumenty}
\usage{
emis_calc(x, y)
}
\arguments{
\item{x}{double}

\item{y}{double}
}
\value{

}
\description{
Funkcja testowa na razie tylko mnozy dwa argumenty
}
\examples{
emis_calc(3,4)
}
